requirements.txt
discord.py==1.7.3
dislash.py==2.0.1